refactor(suite): move Address & Labeling code to @suite packages
