+++ suite/e2e/package.json
@@ -43,6 +43,7 @@
         "@suite/analytics": "workspace:*",
         "@suite/flags": "workspace:*",
         "@suite/intl": "workspace:^",
+        "@suite/labeling": "workspace:*",
         "@suite/metadata": "workspace:*",
         "@suite/router-config": "workspace:*",
         "@suite/settings": "workspace:*",
@@ -52,7 +53,6 @@
         "@trezor/e2e-utils": "workspace:*",
         "@trezor/eslint": "workspace:*",
         "@trezor/protobuf": "workspace:*",
-        "@trezor/suite": "workspace:*",
         "@trezor/suite-data": "workspace:*",
         "@trezor/theme": "workspace:*",
         "@trezor/transport": "workspace:*",
