+++ suite/e2e/tsconfig.json
@@ -50,6 +50,7 @@
         { "path": "../analytics" },
         { "path": "../flags" },
         { "path": "../intl" },
+        { "path": "../labeling" },
         { "path": "../metadata" },
         { "path": "../router-config" },
         { "path": "../settings" },
@@ -61,7 +62,6 @@
         { "path": "../../packages/e2e-utils" },
         { "path": "../../packages/eslint" },
         { "path": "../../packages/protobuf" },
-        { "path": "../../packages/suite" },
         { "path": "../../packages/suite-data" },
         { "path": "../../packages/theme" },
         { "path": "../../packages/transport" },
