+++ suite/labeling/src/LabelingSwitchToLegacyModal.tsx
