+++ suite/labeling/src/index.ts
@@ -1,3 +1,5 @@
 export { Labeling, type LabelingProps } from './Labeling';
+export { LabelingSettings, type LabelingSelectValue } from './LabelingSettings';
+export { LabelingSwitchToLegacyModal } from './LabelingSwitchToLegacyModal';
 export { processLegacyMetadataIntoSuiteSyncThunk } from './processLegacyMetadataIntoSuiteSyncThunk';
 export { selectIsLabelActionEnabled } from './selectIsLabelActionEnabled';
